fix(vision-tools): join_discord_screen actually streams the screen (was a stub)
The first cut (17f71a2) stubbed join_discord_screen to return 'not available'
(wrongly tied to the unrelated discord-voice-view-screen skill). Per Susan: that
skill is NOT what this is about. 'join screen' = stream THIS machine's screen
into the live session — same machinery as start_vision / the 'za warudo screen'
push. Move the tool into vision-tools.ts and make execute call resolveSource('screen')
+ startStream (mirrors startVisionTool, screen source). Now it really works.
Typecheck clean; inlineTools loads it, no dups.
